pyproject.toml
[tool.poetry]
name = "timetable"
version = "0.1.0"
description = ""
authors = ["HansolChoe <hansol614@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
